move data types from helpers to statics
