+++ lib/server/graphql/types/generators/schema-entry-input.js
@@ -7,7 +7,7 @@ import {
   GraphQLInt,
   GraphQLID
 } from 'graphql';
-import {TypesNativeGraphQL} from 'helpers/data-types/native';
+import {TypesNativeGraphQL} from 'statics/data-types/native';
 
 export default (schema) => {
   const propertiesFields = {};
