+++ lib/server/models/schema-entry.js
@@ -1,7 +1,7 @@
 import forEach from 'lodash/forEach';
 import mongoose from 'mongoose';
 import singleStatics from 'statics/schema-single-static-properties';
-import {TypesNative} from 'helpers/data-types/native';
+import {TypesNative} from 'statics/data-types/native';
 
 import SchemaModel from './schema';
