+++ lib/shared/screens/admin/screens/schemas/screens/new/components/builder/progress.jsx
@@ -1,6 +1,6 @@
 import Component from 'components/component';
 import React, {PropTypes} from 'react';
-import {singleFixedProperties} from 'helpers/data-types';
+import {singleFixedProperties} from 'statics/data-types';
 
 import styles from './progress.less';
